Set Plugin Publishing Plugin version in buildSrc with other plugins
This is required to work around Shadow 7 bringing in plugin-publish-plugin
as an implementation dependency which put it on the build's runtime
classpath. It then wasn't possible to set the plugin version in
detekt-gradle-plugin's build script.

This means the version can still be set in detekt's build script until a
corrected Shadow build is released.